OCR/pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "streamlit",
    "groq",
    "python-dotenv",
    "protobuf==3.20.3"
]